# 8355635: Do not collect C strings in C2 scratch buffer